requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.4
django-localflavor==1.1
django-toolbelt==0.0.1
gunicorn==19.3.0
Pillow==2.7.0
psycopg2==2.6
South==1.0.2
static3==0.5.1
